pyproject.toml
[project]
name = "devsecops-env"
version = "0.1.0"
description = "DevSecOps OpenEnv Workspace"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "pydantic>=2.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "uv>=0.1.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
server = "server.app:main"
devsecops-server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server", "models.py", "client.py"]
